+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [v20170625]
+
+  - Address selection widget fixes
+  - Send transaction through Etherscan, refs #21
+  - Android build on Travis CI, refs #20
+  - Run unlock account & send transaction in a thread
+  - Fixes "send" page layout, refs #27
+
 ## [v20170621]
 
   - Travis-CI integration, refs #15